refactor: InputView내 검증 메소드들을 Util화
